v2.9.0 Release

+++ pkg/upload/unity-android.go
@@ -14,7 +14,6 @@ import (
 
 func ProcessUnityAndroid(globalOptions options.CLI, endpoint string, logger log.Logger) error {
 	unityOptions := globalOptions.Upload.UnityAndroid
-	var err error
 	var zipPath string
 	var archList []string
 	var symbolFileList []string
@@ -23,11 +22,7 @@ func ProcessUnityAndroid(globalOptions options.CLI, endpoint string, logger log.
 
 	for _, path := range unityOptions.Path {
 		if utils.IsDir(path) {
-			zipPath, err = utils.FindLatestFileWithSuffix(path, ".symbols.zip")
-
-			if err != nil {
-				return err
-			}
+			zipPath, _ = utils.FindLatestFileWithSuffix(path, ".symbols.zip")
 
 			if aabPath == "" {
 				aabPath, _ = utils.FindLatestFileWithSuffix(path, ".aab")
@@ -80,55 +75,58 @@ func ProcessUnityAndroid(globalOptions options.CLI, endpoint string, logger log.
 		}
 	}
 
-	logger.Debug(fmt.Sprintf("Extracting %s into a temporary directory", filepath.Base(zipPath)))
+	if zipPath == "" {
+		logger.Info("No Unity Android symbols.zip file found, skipping")
+	} else {
+		logger.Debug(fmt.Sprintf("Extracting %s into a temporary directory", filepath.Base(zipPath)))
 
-	if manifestData == nil {
-		manifestData, _ = android.MergeUploadOptionsFromAabManifest("", globalOptions.ApiKey, unityOptions.ApplicationId, unityOptions.BuildUuid, unityOptions.NoBuildUuid, unityOptions.VersionCode, unityOptions.VersionName, logger)
-	}
+		if manifestData == nil {
+			manifestData, _ = android.MergeUploadOptionsFromAabManifest("", globalOptions.ApiKey, unityOptions.ApplicationId, unityOptions.BuildUuid, unityOptions.NoBuildUuid, unityOptions.VersionCode, unityOptions.VersionName, logger)
+		}
 
-	unityDir, err := utils.ExtractFile(zipPath, "unity-android")
+		unityDir, err := utils.ExtractFile(zipPath, "unity-android")
 
-	if err != nil {
-		return err
-	}
-
-	defer os.RemoveAll(unityDir)
-
-	archList, err = utils.BuildDirectoryList([]string{unityDir})
-
-	if err != nil {
-		return err
-	}
-
-	for _, arch := range archList {
-		soPath := filepath.Join(unityDir, arch)
-		fileList, err := utils.BuildFileList([]string{soPath})
 		if err != nil {
 			return err
 		}
-		for _, file := range fileList {
-			if filepath.Base(file) == "libil2cpp.sym.so" && utils.ContainsString(fileList, "libil2cpp.dbg.so") {
-				continue
+
+		defer os.RemoveAll(unityDir)
+
+		archList, err = utils.BuildDirectoryList([]string{unityDir})
+
+		if err != nil {
+			return err
+		}
+
+		for _, arch := range archList {
+			soPath := filepath.Join(unityDir, arch)
+			fileList, err := utils.BuildFileList([]string{soPath})
+			if err != nil {
+				return err
 			}
-			symbolFileList = append(symbolFileList, file)
+			for _, file := range fileList {
+				if filepath.Base(file) == "libil2cpp.sym.so" && utils.ContainsString(fileList, "libil2cpp.dbg.so") {
+					continue
+				}
+				symbolFileList = append(symbolFileList, file)
+			}
+		}
+
+		err = android.UploadAndroidNdk(
+			symbolFileList,
+			manifestData["apiKey"],
+			manifestData["applicationId"],
+			manifestData["versionName"],
+			manifestData["versionCode"],
+			unityOptions.ProjectRoot,
+			endpoint,
+			globalOptions,
+			logger,
+		)
+
+		if err != nil {
+			return err
 		}
 	}
-
-	err = android.UploadAndroidNdk(
-		symbolFileList,
-		manifestData["apiKey"],
-		manifestData["applicationId"],
-		manifestData["versionName"],
-		manifestData["versionCode"],
-		unityOptions.ProjectRoot,
-		endpoint,
-		globalOptions,
-		logger,
-	)
-
-	if err != nil {
-		return err
-	}
-
 	return nil
 }
